refactor: layer nodes are now wrapped in Layer.Node
Layer.Node stores the actual object as a member.
It provides two accessors: obj(layer) and untracked_obj().
The former tracks a read access to the 'obj' param of the node
in the given layer and must therefore be used by analysis engines.
The latter does not track any access.