backtracking: change semantic of ConstraintNotSatisfied
Previously, the CNS exception provided information about the
parameter (i.e. (layer, param, obj)) for which an error occurred.

However, for Check operations, the param is None; for batch
operations, obj might be None.

For generality, we thus applied the following approach:
On any error during an operation, we still insert a new node into
the DecisionGraph. By this, the Graph is sequentialised/treeified
with respect to the operation's dependencies  Then, we fail on
the newly inserted node instead, i.e. the only provided information
by the CNS exception is this Node object.

This significantly simplifies the backtracking as we only need to
find the last revisable decision in the path from the failed operation
to the root of the tree.