MappingEngine: fix incorrect dep cost calc limit
Previously, the costs of a 'dependencies' candidate could not exceed the
number of the 'dependencies' candidates.

For example if the candidate set had length 1 and its only element had
length 2, 'cost' could not be increased by 2 even if both dependencies
of that element could not be satisfied in the same domain.