transform: track implicitly inserted edge nodes
* Layer's _add_node and _add_edge methods do not allow re-insertion of
  the same obj
* non-existent edge nodes are not allowed in Graph.add_edge
* insert_obj does return implicitly inserted edge nodes
* previously inserted nodes are not re-inserted but still returned

This guarantees that the semantic children of source objects are
inserted to the set of associated objects. Now, a node may have an link
to an object of the source layer even if the transformation of that
object did return the node only indirectly via a new edge. As a side
effect, those link parameters may have multiple writers.