framework: check locality of transform operations
We discovered a locality restriction for Transform operations.
A transform must only insert edges between objects whose parents
are already connected. Inserting arbitrary edges becomes problematic
if , e.g., a check operation depends on the outedges of a node. If we
rollback and the re-run operations insert a new edge, the dependencies
of the check operation will change. If, however, we restrict that
objects on the source layer must be connected if they create any edge
between their children, we also restrict the changes that can happen
after a rollback. As the check operation will already depend on the
operations that transform the source objects, a rollback will not change
the dependencies.